requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
langchain
langchain-groq
langchain-community
langgraph
tavily-python
joblib
scikit-learn
